Scrape Jeopardy! Games from J! Archive
This is an initial MVP of the JScraper project, whose ultimate goal is to be able to grab the game contents of an
arbitrary Jeopardy! game so that it can be reproduced for local play. The MVP is able to handle games in which all
clues were revealed and no clues are image or video clues. All three rounds are handled, and the data is loaded into
an easy-to-use model.